Removed geometry option for emacs alias.
I have changed my default screen size for using emacs. I had previously
used a screen size specified by the EMACS_GEOMETRY environment variable.
Because this variable does not work on Mac OS X (I believe) and because
of the larger screen sizes I have available, I have changed to 132x43.

In addition, because of my change to use the customizations specified in
the book, "Clojure for the Brave and True," I now specify this size in
~/.emacs.d/customizations/ui.el (setting the initial-frame-alist
variable).