Installation of emacs 27.1 using scoop
Previous commits **did not** find `emacs` on the path but relied on
adding an alias (on Windows).

This commit removes the alias so that `emacs.exe` because the scoop
package system installs `emacs` in such a way that `emacs.exe` is on the
path (at least when I'm logged in).